requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dropbox==10.1.0
Flask==1.1.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.6
pathspec==0.8.0
psycopg2==2.8.5
python-dotenv==0.12.0
regex==2020.4.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.2.1
